Move exec into function
